// src/types/composer.ts
export enum UserTier {
  FREE = 'free',
  PRO = 'pro',
  ORG = 'org',
}
// ...
export interface TierLimits {
  canCreatePaidEvents: boolean;
  canCreateSeries: boolean;
  canScheduleContent: boolean;
  canLeadSpaces: boolean;
  canPostUnlimitedOpportunities: boolean;
  canAccessFullDIA: boolean;
  canViewCrossCAnalytics: boolean;
  maxOpportunitiesPerMonth: number;
  maxDIASuggestionsPerSession: number;
  maxDIAQueriesPerDay: number;
  maxEventTicketTiers: number;
  maxRSVPQuestions: number;
}